requirements.txt
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.5.0
PyJWT==1.7.1
djoser==2.0.3
django-cors-headers
